+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Fix reset button
+- Fix `f_value/` output directory creation
 - Drop `image` dep to use `png` directly
 - Slim `rfd` dep
 - Bump dependencies